requirements.txt
fastapi
uvicorn[standard]
PyYAML
pydantic
pydantic[email]
psycopg[binary]
psycopg2
SQLAlchemy
pytest
pytest-cov
pytest-notimplemented
coverage
black
mypy
lint
passlib
python-jose[cryptography]

#chroma
chromadb

#langchain
langchain
langchain_community
pypdf

#huggingface
sentence_transformers

#umap
umap-learn